Dropped OLPC support
This commit drops --xembed command line option because it uses
gtk_plug_new() which works correctly only in X11. It is discouraged
to use it in GTK 3.x which supports Wayland.

This commit has been originally writen by Lee Bousfield in 2016,
slightly modified by Rafal Luzynski in 2018. This modified version
does not update the doc directory.