Handle last cross-reference of some x-ref lists.
Sometimes the last cross-ref was thought to be the start of a new
list of cross-references. This commit corrects that by noting
whether the prior cross-ref ended in a semi-colon. If so, then that
indicates that the list continues for at least one more x-ref.

This script processes a file format given to me by Bibles
International. It contained somewhere around 55,000 cross-references.