Attempted fix for scroll_to_insertion_point_on_screen error
The error is that the highlighted verse is not always brought into the
visible area of the editor window, particularly if the verse is later
in a chapter.

This attempts to fix it by 1) splitting highlighting code from
scrolling code and 2) waiting for all gtk pending events to be done
before calling gtk_widget_get_allocation.

This solution seems to help a lot. However, it exposed another bug
that I had before, namely that if I navigate to a verse like Luke 9:50,
the program somehow THEN navigates to Luke 9:0, which is incorrect. I
am at a loss at the moment, but as this is better than it was, I'm
committing it so we can do some further review on it.