Temporarily use cairo_rectangle_int_t.
Since GTK/GDK 3.0 GdkRectangle is explicitly defined as
cairo_rectangle_int_t. However, in 2.x series both structures are
internally the same but are not defined as being the same. The compilers
refuse to convert them implicitly. Therefore we should use
cairo_rectangle_int_t where the documentation says so and not to replace
it with GdkRectangle, and use the explicit conversion when required.

This commit should be reverted once we start using GTK/GDK 3.x series.