Use the GTK3 CSS to style the focused floating window.
Define some CSS classes and assign them to some widgets of the
floating window. Define their style in the CSS file. Define another
class "focused" and its style. Assign the "focused" class to the only
one focused floating window.

This is a port of commit 503b2b4 to GTK3 with CSS.