Rework the system log window into a regular window.
In GTK3 we had some warnings saying "GtkDialog mapped without
a transient parent." In fact, we have never needed this window to be
a modal dialog box. From now it is a regular GtkWindow.

Remove SystemlogDialog::on_system_log_close()

Due to the previous commit this method not only does nothing but
also there is no way to use it ever in the future.